alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-seed==0.3.1
djangorestframework==3.16.1
dotenv==0.9.9
drf-yasg==1.21.11
Faker==38.0.0
idna==3.11
inflection==0.5.1
kombu==5.5.4
mysqlclient==2.2.7
packaging==25.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
